refactor: move SpinalSimAccessSubComponents
